Разобьет редюсер на слайсы в соответствии с предметной областью
