+++ src/components/error/error.tsx
@@ -1,6 +1,6 @@
 import styles from './style.module.css';
 import { useAppSelector } from '../../hooks';
-import { selectError } from '../../store/selectors';
+import { selectError } from '../../store/error/error-selectors';
 
 function Error(): JSX.Element | null {
   const message = useAppSelector(selectError);
