+++ src/components/sorting/sorting.tsx
@@ -2,7 +2,7 @@ import SortingItem from '../sorting-item/sorting-item';
 import { SortType } from '../../const';
 import { useAppSelector } from '../../hooks';
 import { useEffect, useRef, useState } from 'react';
-import { selectSortListType } from '../../store/selectors';
+import { selectSortListType } from '../../store/sorting/sorting-selectors';
 
 function Sorting(): JSX.Element {
   const sortSpanRef = useRef<HTMLElement>(null);
