+++ src/store/error/error-selectors.ts
@@ -1,0 +1,6 @@
+import { NameSpace } from '../../const';
+import { State } from '../../type';
+
+const selectError = (state: State): string | null => state[NameSpace.Error].error;
+
+export {selectError};
