+++ src/store/index.ts
@@ -1,10 +1,11 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { reducer } from './reducer';
 import { createAPI } from '../services/api';
+import rootReducer from './root-reducer';
 
 const api = createAPI();
+
 const store = configureStore({
-  reducer,
+  reducer: rootReducer,
   middleware: (getDefaultMiddleware) =>
     getDefaultMiddleware({
       thunk: {
@@ -13,4 +14,4 @@ const store = configureStore({
     }),
 });
 
-export { store };
+export { store, api };
