+++ src/store/sorting/sorting-selectors.ts
@@ -1,0 +1,6 @@
+import { NameSpace } from '../../const';
+import { State, SortListType } from '../../type';
+
+const selectSortListType = (state: State):SortListType => state[NameSpace.Sorting].currentSortType;
+
+export {selectSortListType};
